Desktop/PROJECTS/TRADE/requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Data Collection
yfinance==0.2.33
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0

# Machine Learning - Core
scikit-learn==1.3.2

# Deep Learning
torch==2.1.2
torchvision==0.16.2

# Technical Indicators
ta==0.11.0

# API & Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Database & Storage
sqlalchemy==2.0.25
pyarrow==14.0.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2

# Date & Time
python-dateutil==2.8.2
